Revert "chore(ci): homeboy autofix — refactor (19 files, 19 fixes)"
This reverts commit a17caa012777cef6db58588da3a10de314ef791f.
