--- src/core/extension/grammar/structural_context.rs
@@ -1,34 +1,0 @@
-//! structural_context — extracted from grammar.rs.
-
-use regex::Regex;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::path::Path;
-use crate::error::{Error, Result};
-use WP_UnitTestCase;
-use DataMachine\Core\Pipeline;
-use super::Region;
-use super::new;
-use super::default;
-use super::super::*;
-
-
-/// Tracks structural context while parsing source text.
-#[derive(Debug, Clone)]
-pub struct StructuralContext {
-    /// Current brace nesting depth.
-    pub depth: i32,
-
-    /// Current region (code, comment, string).
-    pub region: Region,
-
-    /// Stack of block contexts: (kind_label, depth_when_entered).
-    /// Features can push/pop to track impl blocks, test modules, etc.
-    pub block_stack: Vec<(String, i32)>,
-}
-
-impl Default for StructuralContext {
-    fn default() -> Self {
-        Self::new()
-    }
-}
