+++ src/core/extension/grammar_items.rs
@@ -17,9 +17,7 @@
 
 use serde::{Deserialize, Serialize};
 
-use super::grammar::self;
-use super::types::Grammar;
-use super::symbol::Symbol;
+use super::grammar::{self, Grammar, Symbol};
 
 // ============================================================================
 // Types
