+++ src/core/release/resolver.rs
@@ -91,44 +91,3 @@ pub(crate) fn resolve_extensions(
 
     Ok(extensions)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_new_default_path() {
-        let instance = ReleaseCapabilityResolver::default();
-        let extensions = Vec::new();
-        let _result = instance.new(extensions);
-    }
-
-    #[test]
-    fn test_resolve_extensions_if_let_some_configured_component_extensions_as_ref() {
-
-        let result = resolve_extensions();
-        assert!(!result.is_empty(), "expected non-empty collection for: if let Some(configured) = component.extensions.as_ref() {{");
-    }
-
-    #[test]
-    fn test_resolve_extensions_default_path() {
-
-        let result = resolve_extensions();
-        assert!(!result.is_empty(), "expected non-empty collection for: default path");
-    }
-
-    #[test]
-    fn test_resolve_extensions_ok_extensions() {
-
-        let result = resolve_extensions();
-        assert!(!result.is_empty(), "expected non-empty collection for: Ok(extensions)");
-    }
-
-    #[test]
-    fn test_resolve_extensions_has_expected_effects() {
-        // Expected effects: mutation
-
-        let _ = resolve_extensions();
-    }
-
-}
