+++ src/core/release/types/types.rs
@@ -1,5 +1,7 @@
 //! types — extracted from types.rs.
 
+use super::super::*;
+use super::from;
 use crate::engine::pipeline::{self, PipelinePlanStep, PipelineRunResult};
 use crate::is_zero_u32;
 use serde::{Deserialize, Serialize};
