+++ src/core/release/version/version.rs
@@ -177,73 +177,4 @@ mod tests {
             "expected Some for: Some(format!(\"{{}}.{{}}.{{}}\", new_major, new_minor, new_patch))"
         );
     }
-
-    #[test]
-    fn test_parse_version_default_path() {
-        let content = "";
-        let pattern = "";
-        let _result = parse_version(&content, &pattern);
-    }
-
-    #[test]
-    fn test_increment_version_parts_len_3_parts_iter_any_p_p_parse_u32_is_err() {
-        let version = "";
-        let bump_type = "";
-        let result = increment_version(&version, &bump_type);
-        assert!(result.is_none(), "expected None for: parts.len() != 3 || parts.iter().any(|p| p.parse::<u32>().is_err())");
-    }
-
-    #[test]
-    fn test_increment_version_parts_len_3_parts_iter_any_p_p_parse_u32_is_err_2() {
-        let version = "";
-        let bump_type = "";
-        let result = increment_version(&version, &bump_type);
-        assert!(result.is_some(), "expected Some for: parts.len() != 3 || parts.iter().any(|p| p.parse::<u32>().is_err())");
-    }
-
-    #[test]
-    fn test_increment_version_parts_len_3() {
-        let version = "";
-        let bump_type = "";
-        let result = increment_version(&version, &bump_type);
-        assert!(result.is_none(), "expected None for: parts.len() != 3");
-    }
-
-    #[test]
-    fn test_increment_version_parts_len_3_2() {
-        let version = "";
-        let bump_type = "";
-        let _result = increment_version(&version, &bump_type);
-    }
-
-    #[test]
-    fn test_increment_version_parts_len_3_3() {
-        let version = "";
-        let bump_type = "";
-        let _result = increment_version(&version, &bump_type);
-    }
-
-    #[test]
-    fn test_increment_version_default_path() {
-        let version = "";
-        let bump_type = "";
-        let _result = increment_version(&version, &bump_type);
-    }
-
-    #[test]
-    fn test_increment_version_return_none() {
-        let version = "";
-        let bump_type = "";
-        let result = increment_version(&version, &bump_type);
-        assert!(result.is_none(), "expected None for: _ => return None,");
-    }
-
-    #[test]
-    fn test_increment_version_some_format_new_major_new_minor_new_patch() {
-        let version = "";
-        let bump_type = "";
-        let result = increment_version(&version, &bump_type);
-        assert!(result.is_some(), "expected Some for: Some(format!(\"{{}}.{{}}.{{}}\", new_major, new_minor, new_patch))");
-    }
-
 }
